Working on cinematic before main menu
Cinematic scene animation is working good. I didn't plugged it to the main menu yet - i'm working on a UI to display the text when animation is finished, so that user will have to press "spacebar" to start the game. (i will have to think about shortcut to skip it, because I know how boring they can be sometimes... will see next!)